fix: allow existing namespaces to be used
There are some use cases where we'd want to use an existing namespace.
Without this patch, we can get a client-side conflict. If the namespace
already exists, just add the label metadata to the namespace and don't
touch anything else. For any other resource we should keep the existing
behavior.

Signed-off-by: Blake Pettersson <blake.pettersson@gmail.com>